Introduce WorkspaceConfig.
This means we can still have an Action or Closure assigned to the extra property "bndWorkspaceConfigure" and although we can't serialize the Action or Closure into the configuration cache, we can serialize the results of applying it. However, there is a breaking change: the Action or Closure is applied to a WorkspaceConfig instead of a Workspace. We may want to add more methods to WorkspaceConfig for API compatibility.

#6970

Signed-off-by: Eleanor Joslin <ejj@corefiling.com>